SetAuthentication and SetAuthorization has moved to another class. added ICakeLog to log message as setting authentication. Will circle back and fix authorization later.
